refactor(typecheck): split Vue 2 regression modules
