+++ crates/vize_croquis/src/drawer/template/tests.rs
@@ -53,82 +53,6 @@ fn undefined_refs_with_empty_script(template: &str) -> Vec<vize_carton::CompactS
         .collect()
 }
 
-fn undefined_refs_with_empty_legacy_script(template: &str) -> Vec<vize_carton::CompactString> {
-    use vize_armature::parse;
-    use vize_carton::Bump;
-
-    let allocator = Bump::new();
-    let (root, errors) = parse(&allocator, template);
-    assert!(errors.is_empty(), "Template should parse without errors");
-    let mut drawer = Drawer::with_options(DrawerOptions::full()).with_legacy_vue2();
-    drawer.draw_script_plain("export default {}");
-    drawer.draw_template(&root);
-
-    drawer
-        .finish()
-        .undefined_refs
-        .iter()
-        .map(|reference| reference.name.clone())
-        .collect()
-}
-
-#[test]
-fn legacy_slot_scope_attributes_resolve_only_inside_their_subtree() {
-    for attribute in ["slot-scope", "scope"] {
-        let template = format!(
-            r#"<Child><template slot="row" {attribute}="scope"><span>{{{{ scope.row.id }}}}</span></template><p>{{{{ scope }}}}</p></Child>"#,
-        );
-        let undefined = undefined_refs_with_empty_legacy_script(&template);
-        assert_eq!(
-            undefined
-                .iter()
-                .filter(|name| name.as_str() == "scope")
-                .count(),
-            1,
-            "{attribute} must bind scope inside the template without leaking to siblings: {undefined:?}",
-        );
-    }
-}
-
-#[test]
-fn legacy_destructured_slot_scope_records_name_pattern_and_component() {
-    use crate::ScopeData;
-    use vize_armature::parse;
-    use vize_carton::Bump;
-
-    let allocator = Bump::new();
-    let template = r#"<Child><template slot="item" slot-scope="{ row, index }">{{ row.id }} {{ index }}</template></Child>"#;
-    let (root, errors) = parse(&allocator, template);
-    assert!(errors.is_empty(), "Template should parse without errors");
-    let mut drawer = Drawer::with_options(DrawerOptions::full()).with_legacy_vue2();
-    drawer.draw_template(&root);
-    let summary = drawer.finish();
-    let data = summary
-        .scopes
-        .iter()
-        .find_map(|scope| match scope.data() {
-            ScopeData::VSlot(data) => Some(data),
-            _ => None,
-        })
-        .expect("slot-scope must create a v-slot semantic scope");
-
-    assert_eq!(data.name, "item");
-    assert_eq!(data.props_pattern.as_deref(), Some("{ row, index }"));
-    assert_eq!(data.prop_names.as_slice(), ["row", "index"]);
-    assert_eq!(data.component.as_deref(), Some("Child"));
-}
-
-#[test]
-fn legacy_slot_scope_attributes_are_inert_without_legacy_mode() {
-    let undefined = undefined_refs_with_empty_script(
-        r#"<Child><template slot-scope="scope">{{ scope.row.id }}</template></Child>"#,
-    );
-    assert!(
-        undefined.iter().any(|name| name == "scope"),
-        "Vue 3 analysis must not activate Vue 2 attribute sugar: {undefined:?}",
-    );
-}
-
 #[test]
 fn v_scope_keys_resolve_in_subtree() {
     // petite-vue: `count` and `msg` are introduced by v-scope and must not be
